Refactor the code to FixerWallet in recharge and history
